requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
gspread>=5.12.0
google-auth>=2.23.0
python-dotenv>=1.0.0
openai>=1.0.0
google-genai>=0.2.0
weasyprint>=61.0
psutil>=5.9.0
flask>=3.0.0
svix>=1.15.0
pyngrok>=7.0.0
apscheduler>=3.10.0
pytz>=2023.3
dnspython>=2.4.0
playwright>=1.40.0
patchright>=1.0.0
